DF-5 Reading entire test config from environment vars
Previously the tests could not be run locally with the config.json that is
provided with the repository. One would instead need to copy a config given by
Crew Auth directly into the project. This doesn't scale well, nor does it work
with CI, so I've instead added `nconf` and read env vars from `env.json` if the
environment variables are unavailable on the server.

The app will fail to run the tests if the file or environment variables are not
present. env.json is not tracked by version control.